Add empty music player.
